Custom loadouts, new entity database system, draft logo.
Custom loadouts nearly done, new map ent database stuff, 'prototype'
logo thing, possibly a load of debug stuff.

Note on the new map entity system: when storing things in the database,
use a different integer for each entity type (player spawns, map props,
etc.) The data field should contain an integer for anything you want to
specify about the entity (the team of a spawn) you could also use it to
hold a key for another database table if you need to store a lot of
stuff, but try to keep as small as possible. In GM:Initialize there is a
loop that gets all stored ents for the map. Add some conditional
statements to process what the entity is, then do the actual spawning in
a separate function.